+++ TCC/structures.h
@@ -5,12 +5,12 @@
 struct ClientAdjacency {
 	int clientFromId;
 	int clientToId;
-	int travelCost;
+	float travelCost;
 };
 
 struct Client {
 	int id;
-	int demand;
+	float demand;
 	Client() {
 		id = -1;
 		demand = -1;
@@ -24,9 +24,9 @@ struct Client {
 struct Vehicle {
 	int id;
 	int type;
-	int capacity;
-	int cost;
-	int travelCost;
+	float capacity;
+	float cost;
+	float travelCost;
 
 	bool operator==(const Vehicle& other) {
 		return this->id = other.id;
